+++ processor/src/test/fixtures/input/com/example/fragment/includes/provider_cycle/A.java
@@ -1,0 +1,5 @@
+package com.example.fragment.includes.provider_cycle;
+
+@MyFrameworkFragment
+public class A {}
+