+++ processor/src/test/fixtures/input/com/example/fragment/includes/provider_cycle/AImpl.java
@@ -1,0 +1,7 @@
+package com.example.fragment.includes.provider_cycle;
+
+import sting.Fragment;
+
+@Fragment( includes = B.class )
+public interface AImpl {}
+