+++ processor/src/test/fixtures/input/com/example/fragment/includes/provider_cycle/BImpl.java
@@ -1,0 +1,7 @@
+package com.example.fragment.includes.provider_cycle;
+
+import sting.Fragment;
+
+@Fragment( includes = CImpl.class )
+public interface BImpl {}
+