+++ processor/src/test/fixtures/input/com/example/fragment/includes/provider_cycle/MyFrameworkFragment.java
@@ -1,0 +1,9 @@
+package com.example.fragment.includes.provider_cycle;
+
+import sting.StingProvider;
+
+@StingProvider( "[SimpleName]Impl" )
+public @interface MyFrameworkFragment
+{
+}
+